refactor: 가독성을 위해 함수 분리
event 폴더를 만들고 이벤트 진행하는 부분을 모두 event폴더에 모듈로 생성해서 올김
