+++ src/event/triggerDoubleClickEvent.js
@@ -1,0 +1,10 @@
+const editTodoItem = (todoItem) => {
+  console.log(todoItem.closest('label'));
+  todoItem.setAttribute('class', 'editing');
+};
+
+export const triggerDoubleClickEvent = ({ target }) => {
+  if (target.className === 'label') {
+    editTodoItem(target.closest('li'));
+  }
+};
