+++ src/todoList/setTodoItem.js
@@ -1,39 +1,9 @@
-import { countTodoItem } from './countTodoItem.js';
-
-const editTodoItem = (todoItem) => {
-  todoItem.setAttribute('class', 'editing');
-};
-
-const changCompletedTodoItem = (todoItem) => {
-  if (todoItem.className === '') {
-    return todoItem.setAttribute('class', 'completed');
-  }
-  todoItem.removeAttribute('class');
-};
-
-const removeTodoItem = (todoItem) => {
-  todoItem.remove();
-};
-
-const triggerClickEvent = ({ target }) => {
-  if (target.className === 'toggle') {
-    changCompletedTodoItem(target.closest('li'));
-  }
-  if (target.className === 'destroy') {
-    removeTodoItem(target.closest('li'));
-  }
-  countTodoItem();
-};
-
-const triggerDobuleClickEvent = ({ target }) => {
-  if (target.className === 'label') {
-    editTodoItem(target.closest('li'));
-  }
-};
+import { triggerClickEvent } from '../event/triggerClickEvent.js';
+import { triggerDoubleClickEvent } from '../event/triggerDoubleClickEvent.js';
 
 export const setTodoItem = () => {
   const $todoList = document.querySelector('.todo-list');
 
   $todoList.addEventListener('click', triggerClickEvent);
-  $todoList.addEventListener('dblclick', triggerDobuleClickEvent);
+  $todoList.addEventListener('dblclick', triggerDoubleClickEvent);
 };
